Create GlimmerComponent subclass
