+++ packages/ember-htmlbars/lib/main.js
@@ -54,6 +54,7 @@ import legacyEachWithKeywordHelper from 'ember-htmlbars/helpers/-legacy-each-wit
 import htmlSafeHelper from 'ember-htmlbars/helpers/-html-safe';
 import DOMHelper from 'ember-htmlbars/system/dom-helper';
 import Helper, { helper as makeHelper } from 'ember-htmlbars/helper';
+import GlimmerComponent from 'ember-htmlbars/glimmer-component';
 
 // importing adds template bootstrapping
 // initializer to enable embedded templates
@@ -90,6 +91,8 @@ Ember.HTMLBars = {
   registerPlugin: registerPlugin,
   DOMHelper
 };
+
+Ember.GlimmerComponent = GlimmerComponent;
 
 if (isEnabled('ember-htmlbars-helper')) {
   Helper.helper = makeHelper;
